file access done and working as in bash

+++ exec/redirection/heredoc.c
@@ -3,10 +3,10 @@
 /*                                                        :::      ::::::::   */
 /*   heredoc.c                                          :+:      :+:    :+:   */
 /*                                                    +:+ +:+         +:+     */
-/*   By: lomakinavaleria <lomakinavaleria@studen    +#+  +:+       +#+        */
+/*   By: sabdulki <sabdulki@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2024/06/21 14:17:50 by sabdulki          #+#    #+#             */
-/*   Updated: 2024/07/05 15:18:38 by lomakinaval      ###   ########.fr       */
+/*   Updated: 2024/07/05 19:55:49 by sabdulki         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -41,6 +41,8 @@ int	heredoc(int fd, char *file, char *limiter, int mode, t_args *args)
 	// if (check_file_access(file, R_OK))
 	// 	return (-1);
 	fd = get_file_fd(fd, file, mode, HEREDOC);
+	if (fd == -2)
+		return (-2);
 	// fprintf(stderr, "limiter: '%s', its len:  %zu\n", limiter, ft_strlen(limiter));
 	while (1)
 	{
@@ -85,7 +87,7 @@ int	heredoc(int fd, char *file, char *limiter, int mode, t_args *args)
 	return (fd);
 }
 
-void	call_heredocs(char **arr, t_cmd_info *new_cmd, char **limiter_arr, t_args *args)
+int	call_heredocs(char **arr, t_cmd_info *new_cmd, char **limiter_arr, t_args *args)
 {
 	int	i;
 	int	size;
@@ -102,6 +104,8 @@ void	call_heredocs(char **arr, t_cmd_info *new_cmd, char **limiter_arr, t_args *
 		// printf("file: %s, limiter: %s\n", arr[size], limiter);
 		new_cmd->file_read = arr[size];
 		new_cmd->fd_read = heredoc(new_cmd->fd_read, new_cmd->file_read, limiter, new_cmd->mode_read, args);
+		if (new_cmd->fd_read == -2)
+			return (1);
 		if (size != 0)
 		{
 			unlink(arr[size]);
@@ -112,4 +116,7 @@ void	call_heredocs(char **arr, t_cmd_info *new_cmd, char **limiter_arr, t_args *
 	free(arr);
 	free(limiter_arr);
 	new_cmd->fd_read = get_file_fd(new_cmd->fd_read, new_cmd->file_read, new_cmd->mode_read, HEREDOC);
+	if (new_cmd->fd_read == -2)
+		return (1);
+	return (0);
 }
